Refactored layout specific code
