+++ db/migrate/20150725181614_add_image_column_to_user.rb
@@ -1,0 +1,5 @@
+class AddImageColumnToUser < ActiveRecord::Migration
+  def change
+    add_column :users, :image, :string, limit: 1024
+  end
+end
